Add single_output_forward opt. in SingleOutputMha
